Adição de `errorHandler` e validações do `zod` nas rotas e começo do sistema de abas para abrir notas

+++ src/app/api/v1/folders/[folderId]/route.ts
@@ -1,5 +1,8 @@
+import errorHandler from "@/app/errors/errorHandler";
+import { NotFoundError } from "@/app/errors/errors";
 import { PrismaClient } from "@prisma/client";
 import { NextResponse } from "next/server";
+import { z } from "zod";
 
 const prisma = new PrismaClient();
 
@@ -9,56 +12,71 @@ interface IParams {
   };
 }
 
-// Route to edit a Folder
+const folderIdSchema = z.object({
+  folderId: z
+    .string()
+    .regex(/^[0-9]+$/gm, "Only numbers are accepted")
+    .transform((arg) => Number(arg)),
+});
+// Route to edit / rename a Folder
 export async function PUT(req: Request, { params }: IParams) {
-  const { folderId } = await params;
-  const { name } = await req.json();
-
-  const folder = await prisma.folder.update({
-    where: { id: Number(folderId) },
-    data: { name: name },
+  const folderNameSchema = z.object({
+    name: z.string().nonempty(),
   });
 
-  if (!folder) {
+  try {
+    const { folderId } = folderIdSchema.parse(await params);
+    const { name } = folderNameSchema.parse(await req.json());
+
+    // checks if the folder exists
+    const checkFolder = await prisma.folder.findFirst({
+      where: { id: folderId },
+    });
+
+    if (!checkFolder) {
+      throw new NotFoundError({ message: "Couldn't find especified folder." });
+    }
+
+    const folder = await prisma.folder.update({
+      where: { id: Number(folderId) },
+      data: { name: name },
+    });
+
     return NextResponse.json(
-      { message: "Folder not found" },
+      { updatedFolder: folder },
       {
-        status: 404,
+        status: 200,
       },
     );
+  } catch (error) {
+    return errorHandler(error);
   }
-
-  return NextResponse.json(
-    { updatedFolder: folder },
-    {
-      status: 200,
-    },
-  );
 }
 
 // Route to delete a Folder
 export async function DELETE(_req: Request, { params }: IParams) {
-  const { folderId } = await params;
+  try {
+    const { folderId } = folderIdSchema.parse(await params);
 
-  console.log(`FolderID: ${folderId}`);
+    const checkFolder = await prisma.folder.findFirst({
+      where: { id: folderId },
+    });
 
-  const folder = await prisma.folder.delete({
-    where: { id: Number(folderId) },
-  });
+    if (!checkFolder) {
+      throw new NotFoundError({ message: "Couldn't find especified folder." });
+    }
 
-  if (!folder) {
+    const folder = await prisma.folder.delete({
+      where: { id: Number(folderId) },
+    });
+
     return NextResponse.json(
-      { message: "Folder not found" },
+      { deletedFolder: folder },
       {
-        status: 404,
+        status: 200,
       },
     );
+  } catch (error) {
+    return errorHandler(error);
   }
-
-  return NextResponse.json(
-    { deletedFolder: folder },
-    {
-      status: 200,
-    },
-  );
 }
